# frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "utoipa-axum",
		"description": "Utoipa's axum bindings for seamless integration for the two",
		"contact": {
			"name": "Juha Kukkonen",
			"email": "juha7kukkonen@gmail.com"
		},
		"license": {
			"name": "MIT OR Apache-2.0",
			"identifier": "MIT OR Apache-2.0"
		},
		"version": "0.2.0"
	},
	"paths": {
...
		"/attachments/{id}/download": {
			"get": {
				"tags": ["Attachment"],
				"summary": "Get a short-lived download link for a file.",
				"operationId": "download",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Presigned download link",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttachmentDownloadDTO"
								}
							}
						}
					},
					"401": {
						"description": "No auth data found"
					},
					"403": {
						"description": "User has no access to the file's lecture or task"
					},
					"404": {
						"description": "File not found"
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AttachmentDownloadDTO": {
				"type": "object",
				"description": "Short-lived link that downloads the file under its original name.",
				"required": ["url", "expires_in"],
				"properties": {
					"expires_in": {
						"type": "integer",
						"format": "u-int32",
						"description": "Seconds until `url` stops working.",
						"minimum": 0
					},
					"url": {
						"type": "string"
					}
				}
			},
...
		}
	}
}